+++ .changelog/NEXT.md
@@ -2,9 +2,11 @@
 
 ## Features
 
+- **Comms → iMessage is now a Messages sub-nav tab** (same shell as Inbox / Drafts / Sync / Config): `/messages/imessage` with conversation deep links at `/messages/imessage/:chatKey`. Legacy `/imessage` URLs redirect. The page uses the shared Messages header + TabPills instead of a standalone layout.
+
 - **[issue-2415] macOS Contacts sync resolves iMessage names and enriches Tribe.** Settings → Contacts reads AddressBook SQLite (read-only, multi-source) into a machine-local cache. Comms → iMessage conversation titles resolve at read time: **Tribe person name first**, then Contacts (people/companies), then the raw handle. **Fill Tribe phones/emails** matches Contacts to existing people; **Suggest Tribe imports** ranks frequent iMessage counterparts for opt-in add — never dump-all. No write-back to Apple Contacts or Messages.
 
-- **[issue-2413] Comms → iMessage manager for ingested Messages activity.** After Settings → iMessage sync lands events in PortOS, you can browse them under **Comms → iMessage** (`/imessage`): conversation list + short summaries, PortOS-side purge (per event or whole chat), and a handle blocklist that future syncs honor. Deletes never write Apple's `chat.db` (read-only boundary unchanged). Settings → iMessage links to the manager and surfaces when more history remains to drain.
+- **[issue-2413] Comms → iMessage manager for ingested Messages activity.** After Settings → iMessage sync lands events in PortOS, you can browse them under **Comms → iMessage** (`/messages/imessage`): conversation list + short summaries, PortOS-side purge (per event or whole chat), and a handle blocklist that future syncs honor. Deletes never write Apple's `chat.db` (read-only boundary unchanged). Settings → iMessage links to the manager and surfaces when more history remains to drain.
 
 - **A never-run cron task no longer fires an immediate "catch-up" for a scheduled slot that predates when the task was set up.** `shouldRunTask`'s cron catch-up bounded a never-run task only by "the occurrence before the most recent slot" — which is always earlier than that slot, so the guard was vacuous and *every* never-run cron task reported `cron-catch-up` for its most recent past slot. For long periods this was jarring: a weekly `branch-cleanup` scheduled `0 9 * * 0` (Sunday 09:00) that was enabled mid-week immediately showed as **due now (catch-up)** on the Schedule Timeline for *last* Sunday's slot — a slot that elapsed before the task existed and was never actually missed. Tasks now carry a `createdAt` timestamp (stamped when first configured; `loadSchedule` backfills existing installs to "now", which only suppresses catch-up for already-past slots and never affects future cadence), and never-run catch-up only fires for a slot that elapsed *after* `createdAt`. So a weekly task enabled on a Friday now waits for the coming Sunday instead of running immediately, while a task that was genuinely configured before a missed slot (daemon down / check window missed) still catches up as before.
