+++ client/src/components/settings/IMessageTab.jsx
@@ -97,7 +97,7 @@ export function IMessageTab() {
           and feeds both Tribe touchpoints and the activity timeline. Machine-local — nothing federates to peers.
           Requires macOS <strong>Full Disk Access</strong> for the process running PortOS.
           Browse, purge spam, and manage ingested events under{' '}
-          <Link to="/imessage" className="text-port-accent hover:underline inline-flex items-center gap-1">
+          <Link to="/messages/imessage" className="text-port-accent hover:underline inline-flex items-center gap-1">
             Comms → iMessage <ExternalLink size={12} />
           </Link>
           {' '}— deletes there remove PortOS copies only, never Apple Messages.
@@ -189,7 +189,7 @@ export function IMessageTab() {
           <div className="flex flex-wrap items-center justify-between gap-2 mb-2">
             <h3 className="text-sm font-semibold text-white">Last sync</h3>
             <Link
-              to="/imessage"
+              to="/messages/imessage"
               className="inline-flex items-center gap-1 text-xs text-port-accent hover:underline"
             >
               Open iMessage manager <ExternalLink size={12} />
@@ -209,7 +209,7 @@ export function IMessageTab() {
               First batches are often older messages; open{' '}
               <Link to="/timeline" className="text-port-accent hover:underline">Timeline</Link>
               {' '}on those dates or use{' '}
-              <Link to="/imessage" className="text-port-accent hover:underline">Comms → iMessage</Link>
+              <Link to="/messages/imessage" className="text-port-accent hover:underline">Comms → iMessage</Link>
               {' '}to browse by conversation.
             </p>
           )}
