+++ client/src/pages/Messages.jsx
@@ -1,5 +1,5 @@
-import { useNavigate } from 'react-router-dom';
-import { Mail, RefreshCw, Settings } from 'lucide-react';
+import { useNavigate, useParams } from 'react-router-dom';
+import { Mail, RefreshCw, Settings, MessageSquare } from 'lucide-react';
 import { useState, useEffect, useCallback } from 'react';
 import * as api from '../services/api';
 import BrailleSpinner from '../components/BrailleSpinner';
@@ -11,18 +11,25 @@ import InboxTab from '../components/messages/InboxTab';
 import ConfigTab from '../components/messages/ConfigTab';
 import DraftsTab from '../components/messages/DraftsTab';
 import SyncTab from '../components/messages/SyncTab';
+import IMessageTab from '../components/messages/IMessageTab';
 
 // Exported for the nav-manifest tab-coverage guard (server/lib/navManifest.test.js).
+// `fullBleed: true` — tab owns internal scroll/height; Messages skips padded overflow wrapper.
 export const TABS = [
   { id: 'inbox', label: 'Inbox', icon: Mail },
   { id: 'drafts', label: 'Drafts', icon: Mail },
+  { id: 'imessage', label: 'iMessage', icon: MessageSquare, fullBleed: true },
   { id: 'sync', label: 'Sync', icon: RefreshCw },
-  { id: 'config', label: 'Config', icon: Settings }
+  { id: 'config', label: 'Config', icon: Settings },
 ];
+
+const FULL_BLEED_TAB_IDS = new Set(TABS.filter((t) => t.fullBleed).map((t) => t.id));
 
 export default function Messages() {
   const navigate = useNavigate();
+  const { chatKey } = useParams();
   const activeTab = useValidTab(TABS, 'inbox');
+  const fullBleed = FULL_BLEED_TAB_IDS.has(activeTab);
   const [accounts, setAccounts] = useState([]);
   const [loading, setLoading] = useState(true);
 
@@ -35,6 +42,14 @@ export default function Messages() {
   useEffect(() => {
     fetchAccounts();
   }, [fetchAccounts]);
+
+  // Deep-link cleanup: only the imessage tab uses :chatKey. Drop a stale second
+  // segment if the user lands on e.g. /messages/inbox/<something>.
+  useEffect(() => {
+    if (chatKey && activeTab !== 'imessage') {
+      navigate(`/messages/${activeTab}`, { replace: true });
+    }
+  }, [chatKey, activeTab, navigate]);
 
   const handleTabChange = (tabId) => {
     navigate(`/messages/${tabId}`);
@@ -50,6 +65,8 @@ export default function Messages() {
         return <DraftsTab accounts={accounts} />;
       case 'sync':
         return <SyncTab accounts={accounts} onRefresh={fetchAccounts} />;
+      case 'imessage':
+        return <IMessageTab />;
       default:
         return <InboxTab accounts={accounts} />;
     }
@@ -64,7 +81,7 @@ export default function Messages() {
   }
 
   return (
-    <div className="flex flex-col h-full">
+    <div className="flex flex-col h-full min-h-0">
       <PageHeader
         icon={Mail}
         title="Messages"
@@ -74,7 +91,7 @@ export default function Messages() {
 
       <TabPills tabs={TABS} activeTab={activeTab} onChange={handleTabChange} ariaLabel="Messages sections" />
 
-      <div className="flex-1 overflow-auto p-4">
+      <div className={`flex-1 min-h-0 ${fullBleed ? 'overflow-hidden' : 'overflow-auto p-4'}`}>
         {renderTabContent()}
       </div>
     </div>
